+++ src/Helpers.php
@@ -3,25 +3,13 @@
 declare(strict_types=1);
 use Crunz\Utils;
 
-if (!\function_exists('split_camel')) {
-    /**
-     * Split camel case words to a sentence.
-     *
-     * @param string $text
-     *
-     * @return string
-     */
-    function split_camel($text)
-    {
-        return Utils::splitCamel($text);
-    }
-}
-
 if (!\function_exists('word2number')) {
     /**
      * Convert words to numbers.
      *
      * @param string $text
+     *
+     * @internal
      *
      * @return string
      */
@@ -39,6 +27,8 @@ if (!\function_exists('array_only')) {
      * @param array        $array
      * @param array|string $keys
      *
+     * @internal
+     *
      * @return array
      */
     function array_only($array, $keys)
@@ -53,6 +43,8 @@ if (!\function_exists('setbase')) {
      *
      * @param string $dir
      *
+     * @internal
+     *
      * @return string
      */
     function setbase($dir)
@@ -64,6 +56,8 @@ if (!\function_exists('setbase')) {
 if (!\function_exists('getbase')) {
     /**
      * Return the project's root directory.
+     *
+     * @internal
      *
      * @return string
      */
@@ -79,6 +73,8 @@ if (!\function_exists('generate_path')) {
      *
      * @param string $relative_path
      *
+     * @internal
+     *
      * @return string
      */
     function generate_path($relative_path)
@@ -92,6 +88,8 @@ if (!\function_exists('getroot')) {
      * Return absolute path for relative path.
      *
      * @param string $autoloader
+     *
+     * @internal
      *
      * @return string
      */