+++ src/Utils.php
@@ -4,27 +4,14 @@ declare(strict_types=1);
 
 namespace Crunz;
 
-class Utils
+final class Utils
 {
-    /**
-     * Split camel case string.
-     *
-     * @param string $text
-     *
-     * @return string
-     */
-    public static function splitCamel($text)
-    {
-        $pattern = '/(?<=[a-z])(?=[A-Z])/x';
-        $segments = \preg_split($pattern, $text);
-
-        return \mb_strtolower(\implode($segments, ' '));
-    }
-
     /**
      * Set project's root directory.
      *
      * @param string $autoloader
+     *
+     * @internal
      *
      * @return string
      */
@@ -35,6 +22,8 @@ class Utils
 
     /**
      * Return project's root directory.
+     *
+     * @internal
      *
      * @return string
      */
@@ -48,6 +37,8 @@ class Utils
      *
      * @param string $autoloader
      *
+     * @internal
+     *
      * @return string
      */
     public static function getRoot($autoloader)
@@ -59,6 +50,8 @@ class Utils
      * return absolute path for relative path.
      *
      * @param string $relative_path
+     *
+     * @internal
      *
      * @return string
      */
@@ -74,6 +67,8 @@ class Utils
      * @param array        $array
      * @param array|string $keys
      *
+     * @internal
+     *
      * @return array
      */
     public static function arrayOnly($array, $keys)
@@ -85,6 +80,8 @@ class Utils
      * Converts words to numbers.
      *
      * @param string $text
+     *
+     * @internal
      *
      * @return string
      */