+++ tests/Unit/UtilsTest.php
@@ -9,13 +9,6 @@ use PHPUnit\Framework\TestCase;
 
 class UtilsTest extends TestCase
 {
-    public function testSplitCamel()
-    {
-        $this->assertEquals('thirty seven', Utils::splitCamel('thirtySeven'));
-        $this->assertEquals('one thousand and five hundred sixty three', Utils::splitCamel('oneThousandAndFiveHundredSixtyThree'));
-        $this->assertEquals('seven hundred eighty five', Utils::splitCamel('sevenHundredEightyFive'));
-    }
-
     public function testWordToNumber()
     {
         $this->assertEquals(65, Utils::wordToNumber('sixty five'));